Web App/requirements.txt
streamlit
transformers
beautifulsoup4
requests
youtube-dl
PyPDF2
python-docx